Struct rusoto_sns::SnsClient

pub struct SnsClient { /* fields omitted */ }

A client for the Amazon SNS API.

Methods

impl SnsClient

pub fn new(region: Region) -> SnsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SnsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,