Trait rusoto_sns::Sns

pub trait Sns {
    fn add_permission(
        &self, 
        input: AddPermissionInput
    ) -> RusotoFuture<(), AddPermissionError>;

    fn check_if_phone_number_is_opted_out(
        &self, 
        input: CheckIfPhoneNumberIsOptedOutInput
    ) -> RusotoFuture<CheckIfPhoneNumberIsOptedOutResponse, CheckIfPhoneNumberIsOptedOutError>;

    fn confirm_subscription(
        &self, 
        input: ConfirmSubscriptionInput
    ) -> RusotoFuture<ConfirmSubscriptionResponse, ConfirmSubscriptionError>;

    fn create_platform_application(
        &self, 
        input: CreatePlatformApplicationInput
    ) -> RusotoFuture<CreatePlatformApplicationResponse, CreatePlatformApplicationError>;

    fn create_platform_endpoint(
        &self, 
        input: CreatePlatformEndpointInput
    ) -> RusotoFuture<CreateEndpointResponse, CreatePlatformEndpointError>;

    fn create_topic(
        &self, 
        input: CreateTopicInput
    ) -> RusotoFuture<CreateTopicResponse, CreateTopicError>;

    fn delete_endpoint(
        &self, 
        input: DeleteEndpointInput
    ) -> RusotoFuture<(), DeleteEndpointError>;

    fn delete_platform_application(
        &self, 
        input: DeletePlatformApplicationInput
    ) -> RusotoFuture<(), DeletePlatformApplicationError>;

    fn delete_topic(
        &self, 
        input: DeleteTopicInput
    ) -> RusotoFuture<(), DeleteTopicError>;

    fn get_endpoint_attributes(
        &self, 
        input: GetEndpointAttributesInput
    ) -> RusotoFuture<GetEndpointAttributesResponse, GetEndpointAttributesError>;

    fn get_platform_application_attributes(
        &self, 
        input: GetPlatformApplicationAttributesInput
    ) -> RusotoFuture<GetPlatformApplicationAttributesResponse, GetPlatformApplicationAttributesError>;

    fn get_sms_attributes(
        &self, 
        input: GetSMSAttributesInput
    ) -> RusotoFuture<GetSMSAttributesResponse, GetSMSAttributesError>;

    fn get_subscription_attributes(
        &self, 
        input: GetSubscriptionAttributesInput
    ) -> RusotoFuture<GetSubscriptionAttributesResponse, GetSubscriptionAttributesError>;

    fn get_topic_attributes(
        &self, 
        input: GetTopicAttributesInput
    ) -> RusotoFuture<GetTopicAttributesResponse, GetTopicAttributesError>;

    fn list_endpoints_by_platform_application(
        &self, 
        input: ListEndpointsByPlatformApplicationInput
    ) -> RusotoFuture<ListEndpointsByPlatformApplicationResponse, ListEndpointsByPlatformApplicationError>;

    fn list_phone_numbers_opted_out(
        &self, 
        input: ListPhoneNumbersOptedOutInput
    ) -> RusotoFuture<ListPhoneNumbersOptedOutResponse, ListPhoneNumbersOptedOutError>;

    fn list_platform_applications(
        &self, 
        input: ListPlatformApplicationsInput
    ) -> RusotoFuture<ListPlatformApplicationsResponse, ListPlatformApplicationsError>;

    fn list_subscriptions(
        &self, 
        input: ListSubscriptionsInput
    ) -> RusotoFuture<ListSubscriptionsResponse, ListSubscriptionsError>;

    fn list_subscriptions_by_topic(
        &self, 
        input: ListSubscriptionsByTopicInput
    ) -> RusotoFuture<ListSubscriptionsByTopicResponse, ListSubscriptionsByTopicError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn list_topics(
        &self, 
        input: ListTopicsInput
    ) -> RusotoFuture<ListTopicsResponse, ListTopicsError>;

    fn opt_in_phone_number(
        &self, 
        input: OptInPhoneNumberInput
    ) -> RusotoFuture<OptInPhoneNumberResponse, OptInPhoneNumberError>;

    fn publish(
        &self, 
        input: PublishInput
    ) -> RusotoFuture<PublishResponse, PublishError>;

    fn remove_permission(
        &self, 
        input: RemovePermissionInput
    ) -> RusotoFuture<(), RemovePermissionError>;

    fn set_endpoint_attributes(
        &self, 
        input: SetEndpointAttributesInput
    ) -> RusotoFuture<(), SetEndpointAttributesError>;

    fn set_platform_application_attributes(
        &self, 
        input: SetPlatformApplicationAttributesInput
    ) -> RusotoFuture<(), SetPlatformApplicationAttributesError>;

    fn set_sms_attributes(
        &self, 
        input: SetSMSAttributesInput
    ) -> RusotoFuture<SetSMSAttributesResponse, SetSMSAttributesError>;

    fn set_subscription_attributes(
        &self, 
        input: SetSubscriptionAttributesInput
    ) -> RusotoFuture<(), SetSubscriptionAttributesError>;

    fn set_topic_attributes(
        &self, 
        input: SetTopicAttributesInput
    ) -> RusotoFuture<(), SetTopicAttributesError>;

    fn subscribe(
        &self, 
        input: SubscribeInput
    ) -> RusotoFuture<SubscribeResponse, SubscribeError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn unsubscribe(
        &self, 
        input: UnsubscribeInput
    ) -> RusotoFuture<(), UnsubscribeError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;
}

Trait representing the capabilities of the Amazon SNS API. Amazon SNS clients implement this trait.

Required methods

fn add_permission(
&self,
input: AddPermissionInput
) -> RusotoFuture<(), AddPermissionError>

Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions.

fn check_if_phone_number_is_opted_out(
&self,
input: CheckIfPhoneNumberIsOptedOutInput
) -> RusotoFuture<CheckIfPhoneNumberIsOptedOutResponse, CheckIfPhoneNumberIsOptedOutError>

Accepts a phone number and indicates whether the phone holder has opted out of receiving SMS messages from your account. You cannot send SMS messages to a number that is opted out.

To resume sending messages, you can opt in the number by using the OptInPhoneNumber action.

fn confirm_subscription(
&self,
input: ConfirmSubscriptionInput
) -> RusotoFuture<ConfirmSubscriptionResponse, ConfirmSubscriptionError>

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true".

fn create_platform_application(
&self,
input: CreatePlatformApplicationInput
) -> RusotoFuture<CreatePlatformApplicationResponse, CreatePlatformApplicationError>

Creates a platform application object for one of the supported push notification services, such as APNS and FCM, to which devices and mobile apps may register. You must specify PlatformPrincipal and PlatformCredential attributes when using the CreatePlatformApplication action. The PlatformPrincipal is received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". The PlatformCredential is also received from the notification service. For WNS, PlatformPrincipal is "Package Security Identifier". For MPNS, PlatformPrincipal is "TLS certificate". For Baidu, PlatformPrincipal is "API key".

For APNS/APNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". For WNS, PlatformCredential is "secret key". For MPNS, PlatformCredential is "private key". For Baidu, PlatformCredential is "secret key". The PlatformApplicationArn that is returned when using CreatePlatformApplication is then used as an attribute for the CreatePlatformEndpoint action. For more information, see Using Amazon SNS Mobile Push Notifications. For more information about obtaining the PlatformPrincipal and PlatformCredential for each of the supported push notification services, see Getting Started with Apple Push Notification Service, Getting Started with Amazon Device Messaging, Getting Started with Baidu Cloud Push, Getting Started with Google Cloud Messaging for Android, Getting Started with MPNS, or Getting Started with WNS.

fn create_platform_endpoint(
&self,
input: CreatePlatformEndpointInput
) -> RusotoFuture<CreateEndpointResponse, CreatePlatformEndpointError>

Creates an endpoint for a device and mobile app on one of the supported push notification services, such as GCM and APNS. CreatePlatformEndpoint requires the PlatformApplicationArn that is returned from CreatePlatformApplication. The EndpointArn that is returned when using CreatePlatformEndpoint can then be used by the Publish action to send a message to a mobile app or by the Subscribe action for subscription to a topic. The CreatePlatformEndpoint action is idempotent, so if the requester already owns an endpoint with the same device token and attributes, that endpoint's ARN is returned without creating a new endpoint. For more information, see Using Amazon SNS Mobile Push Notifications.

When using CreatePlatformEndpoint with Baidu, two attributes must be provided: ChannelId and UserId. The token field must also contain the ChannelId. For more information, see Creating an Amazon SNS Endpoint for Baidu.

fn create_topic(
&self,
input: CreateTopicInput
) -> RusotoFuture<CreateTopicResponse, CreateTopicError>

Creates a topic to which notifications can be published. Users can create at most 100,000 topics. For more information, see https://aws.amazon.com/sns. This action is idempotent, so if the requester already owns a topic with the specified name, that topic's ARN is returned without creating a new topic.

fn delete_endpoint(
&self,
input: DeleteEndpointInput
) -> RusotoFuture<(), DeleteEndpointError>

Deletes the endpoint for a device and mobile app from Amazon SNS. This action is idempotent. For more information, see Using Amazon SNS Mobile Push Notifications.

When you delete an endpoint that is also subscribed to a topic, then you must also unsubscribe the endpoint from the topic.

fn delete_platform_application(
&self,
input: DeletePlatformApplicationInput
) -> RusotoFuture<(), DeletePlatformApplicationError>

Deletes a platform application object for one of the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications.

fn delete_topic(
&self,
input: DeleteTopicInput
) -> RusotoFuture<(), DeleteTopicError>

Deletes a topic and all its subscriptions. Deleting a topic might prevent some messages previously sent to the topic from being delivered to subscribers. This action is idempotent, so deleting a topic that does not exist does not result in an error.

fn get_endpoint_attributes(
&self,
input: GetEndpointAttributesInput
) -> RusotoFuture<GetEndpointAttributesResponse, GetEndpointAttributesError>

Retrieves the endpoint attributes for a device on one of the supported push notification services, such as GCM and APNS. For more information, see Using Amazon SNS Mobile Push Notifications.

fn get_platform_application_attributes(
&self,
input: GetPlatformApplicationAttributesInput
) -> RusotoFuture<GetPlatformApplicationAttributesResponse, GetPlatformApplicationAttributesError>

Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications.

fn get_sms_attributes(
&self,
input: GetSMSAttributesInput
) -> RusotoFuture<GetSMSAttributesResponse, GetSMSAttributesError>

Returns the settings for sending SMS messages from your account.

These settings are set with the SetSMSAttributes action.

fn get_subscription_attributes(
&self,
input: GetSubscriptionAttributesInput
) -> RusotoFuture<GetSubscriptionAttributesResponse, GetSubscriptionAttributesError>

Returns all of the properties of a subscription.

fn get_topic_attributes(
&self,
input: GetTopicAttributesInput
) -> RusotoFuture<GetTopicAttributesResponse, GetTopicAttributesError>

Returns all of the properties of a topic. Topic properties returned might differ based on the authorization of the user.

fn list_endpoints_by_platform_application(
&self,
input: ListEndpointsByPlatformApplicationInput
) -> RusotoFuture<ListEndpointsByPlatformApplicationResponse, ListEndpointsByPlatformApplicationError>

Lists the endpoints and endpoint attributes for devices in a supported push notification service, such as GCM and APNS. The results for ListEndpointsByPlatformApplication are paginated and return a limited list of endpoints, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListEndpointsByPlatformApplication again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

This action is throttled at 30 transactions per second (TPS).

fn list_phone_numbers_opted_out(
&self,
input: ListPhoneNumbersOptedOutInput
) -> RusotoFuture<ListPhoneNumbersOptedOutResponse, ListPhoneNumbersOptedOutError>

Returns a list of phone numbers that are opted out, meaning you cannot send SMS messages to them.

The results for ListPhoneNumbersOptedOut are paginated, and each page returns up to 100 phone numbers. If additional phone numbers are available after the first page of results, then a NextToken string will be returned. To receive the next page, you call ListPhoneNumbersOptedOut again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null.

fn list_platform_applications(
&self,
input: ListPlatformApplicationsInput
) -> RusotoFuture<ListPlatformApplicationsResponse, ListPlatformApplicationsError>

Lists the platform application objects for the supported push notification services, such as APNS and GCM. The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

This action is throttled at 15 transactions per second (TPS).

fn list_subscriptions(
&self,
input: ListSubscriptionsInput
) -> RusotoFuture<ListSubscriptionsResponse, ListSubscriptionsError>

Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptions call to get further results.

This action is throttled at 30 transactions per second (TPS).

fn list_subscriptions_by_topic(
&self,
input: ListSubscriptionsByTopicInput
) -> RusotoFuture<ListSubscriptionsByTopicResponse, ListSubscriptionsByTopicError>

Returns a list of the subscriptions to a specific topic. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptionsByTopic call to get further results.

This action is throttled at 30 transactions per second (TPS).

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

List all tags added to the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the Amazon Simple Notification Service Developer Guide.

fn list_topics(
&self,
input: ListTopicsInput
) -> RusotoFuture<ListTopicsResponse, ListTopicsError>

Returns a list of the requester's topics. Each call returns a limited list of topics, up to 100. If there are more topics, a NextToken is also returned. Use the NextToken parameter in a new ListTopics call to get further results.

This action is throttled at 30 transactions per second (TPS).

fn opt_in_phone_number(
&self,
input: OptInPhoneNumberInput
) -> RusotoFuture<OptInPhoneNumberResponse, OptInPhoneNumberError>

Use this request to opt in a phone number that is opted out, which enables you to resume sending SMS messages to the number.

You can opt in a phone number only once every 30 days.

fn publish(
&self,
input: PublishInput
) -> RusotoFuture<PublishResponse, PublishError>

Sends a message to an Amazon SNS topic or sends a text message (SMS message) directly to a phone number.

If you send a message to a topic, Amazon SNS delivers the message to each endpoint that is subscribed to the topic. The format of the message depends on the notification protocol for each subscribed endpoint.

When a messageId is returned, the message has been saved and Amazon SNS will attempt to deliver it shortly.

To use the Publish action for sending a message to a mobile endpoint, such as an app on a Kindle device or mobile phone, you must specify the EndpointArn for the TargetArn parameter. The EndpointArn is returned when making a call with the CreatePlatformEndpoint action.

For more information about formatting messages, see Send Custom Platform-Specific Payloads in Messages to Mobile Devices.

fn remove_permission(
&self,
input: RemovePermissionInput
) -> RusotoFuture<(), RemovePermissionError>

Removes a statement from a topic's access control policy.

fn set_endpoint_attributes(
&self,
input: SetEndpointAttributesInput
) -> RusotoFuture<(), SetEndpointAttributesError>

Sets the attributes for an endpoint for a device on one of the supported push notification services, such as GCM and APNS. For more information, see Using Amazon SNS Mobile Push Notifications.

fn set_platform_application_attributes(
&self,
input: SetPlatformApplicationAttributesInput
) -> RusotoFuture<(), SetPlatformApplicationAttributesError>

Sets the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications. For information on configuring attributes for message delivery status, see Using Amazon SNS Application Attributes for Message Delivery Status.

fn set_sms_attributes(
&self,
input: SetSMSAttributesInput
) -> RusotoFuture<SetSMSAttributesResponse, SetSMSAttributesError>

Use this request to set the default settings for sending SMS messages and receiving daily SMS usage reports.

You can override some of these settings for a single message when you use the Publish action with the MessageAttributes.entry.N parameter. For more information, see Sending an SMS Message in the Amazon SNS Developer Guide.

fn set_subscription_attributes(
&self,
input: SetSubscriptionAttributesInput
) -> RusotoFuture<(), SetSubscriptionAttributesError>

Allows a subscription owner to set an attribute of the subscription to a new value.

fn set_topic_attributes(
&self,
input: SetTopicAttributesInput
) -> RusotoFuture<(), SetTopicAttributesError>

Allows a topic owner to set an attribute of the topic to a new value.

fn subscribe(
&self,
input: SubscribeInput
) -> RusotoFuture<SubscribeResponse, SubscribeError>

Prepares to subscribe an endpoint by sending the endpoint a confirmation message. To actually create a subscription, the endpoint owner must call the ConfirmSubscription action with the token from the confirmation message. Confirmation tokens are valid for three days.

This action is throttled at 100 transactions per second (TPS).

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Add tags to the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the Amazon SNS Developer Guide.

When you use topic tags, keep the following guidelines in mind:

Adding more than 50 tags to a topic isn't recommended.

Tags don't have any semantic meaning. Amazon SNS interprets tags as character strings.

Tags are case-sensitive.

A new tag with a key identical to that of an existing tag overwrites the existing tag.

Tagging actions are limited to 10 TPS per AWS account. If your application requires a higher throughput, file a technical support request.

For a full list of tag restrictions, see Limits Related to Topics in the Amazon SNS Developer Guide.

fn unsubscribe(
&self,
input: UnsubscribeInput
) -> RusotoFuture<(), UnsubscribeError>

Deletes a subscription. If the subscription requires authentication for deletion, only the owner of the subscription or the topic's owner can unsubscribe, and an AWS signature is required. If the Unsubscribe call does not require authentication and the requester is not the subscription owner, a final cancellation message is delivered to the endpoint, so that the endpoint owner can easily resubscribe to the topic if the Unsubscribe request was unintended.

This action is throttled at 100 transactions per second (TPS).

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Remove tags from the specified Amazon SNS topic. For an overview, see Amazon SNS Tags in the Amazon SNS Developer Guide.

Loading content...